Update and correct information in thesis template files.
This commit includes several changes to the novathesis template files, improving the overall correctness. Notably, it fixed typographical errors in the titles and author's name and added academic titles to advisers in the cover file. Furthermore, it commented out a timing function in the template. The bibliography has also been cleaned up, with several article records updated or removed for clarity.