Add new article to zotero bibliography
A new article titled "Using weather derivatives to hedge precipitation exposure" was added to the zotero.bib file in the novathesis_template repo. This addition expands the library of resources for subsequent research and reference. The newly added article offers important insights on the use of weather derivatives for risk management, enhancing the hedging strategies discussion.