Update references in novathesis_template's Bibliography
Adjusted the order of keywords in multiple bibliographic entries to prioritize main subjects. Added multiple new bibliographic entries to illustrate the application of financial mathematics in renewable energy markets and the advances in quantitative risk management of wind power and other energy commodities. This update broadens the scope of literature