Update bibliography and modify document files
A new entry was added to the bibliography file and changes were made to the novathesis configuration and several chapters. The novathesis document status has been updated to reflect a more current stage of the project. Changes to chapters include text modifications and additions to improve clarity as well as graph updates to better reflect research findings.