Updated thesis chapter on wind modelling
In this commit, text revisions for clarity and grammar were implemented and a new subsection regarding state components was introduced into the wind modelling chapter of the thesis. This new content primarily comprises a detailed discussion on the trend component of a structural time series model, including its definition and mathematical formulation.