Refactor wind prediction function and handle NaN values
The function `get_wind_prev_pivot` has been refactored to now accept a tuple of zone names and a history number of years as input. It also now includes a handling line for NaN values within the dataset, substituting them with 0 - a necessity for projects that started later than the historical data retrieval timeframe.