Add R graphics settings and package cache files
This commit introduces two new .xml files related to the settings of the R environment. The rGraphicsSettings.xml defines the graphics settings, such as the height, resolution, version, and width. The rAvailablePackageCache.xml stores the cache of available R packages from a specified repository, aiding in quicker package retrieval and installation.